pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "insta-scraper"
version = "0.1.0"
description = "Instagram public profile scraping service (Playwright + FastAPI)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.7.0",
  "playwright>=1.52.0",
  "requests>=2.32.0",
  "python-dateutil>=2.9.0",
  "tzdata>=2024.1",
  "openpyxl>=3.1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["data*", "exports*", "tests*"]
